fix(controller): guard DS4 output path and cast JNI scalars
- Dualshock4Controller.sendCommand() now synchronizes on the shared
  outputLock and re-checks outputClosed under it, so a queued rumble
  cannot follow stop()'s final clear (same race as DualSense).
- BridgeClSetAdaptiveTriggers casts scalar args to jshort/jbyte before
  the JNI call, as BridgeClRumble already documents for CheckJNI.

+++ app/src/main/jni/moonlight-core/callbacks.c
@@ -483,8 +483,10 @@ void BridgeClSetAdaptiveTriggers(uint16_t controllerNumber, uint8_t eventFlags,
     (*env)->SetByteArrayRegion(env, leftArray, 0, DS_EFFECT_PAYLOAD_SIZE, (const jbyte*)left);
     (*env)->SetByteArrayRegion(env, rightArray, 0, DS_EFFECT_PAYLOAD_SIZE, (const jbyte*)right);
     if (!(*env)->ExceptionCheck(env)) {
+        // Casts to signed types are required for CheckJNI; see BridgeClRumble.
         (*env)->CallStaticVoidMethod(env, GlobalBridgeClass, BridgeClSetAdaptiveTriggersMethod,
-                                     controllerNumber, eventFlags, typeLeft, typeRight,
+                                     (jshort)controllerNumber, (jbyte)eventFlags,
+                                     (jbyte)typeLeft, (jbyte)typeRight,
                                      leftArray, rightArray);
     }
